Conflicts:
	lib/termtter/client.rb

+++ lib/plugins/defaults/list.rb
@@ -22,6 +22,7 @@ module Termtter::Client
         options = {}
       end
 
+      last_error = nil
       if arg.empty?
         event = :list_friends_timeline
         statuses = Termtter::API.twitter.home_timeline(options)
@@ -33,14 +34,28 @@ module Termtter::Client
             user_name, slug = *user.split('/')
             user_name = config.user_name if user_name.empty?
             user_name = normalize_as_user_name(user_name)
+            options[:per_page] = options[:count]
+            options.delete(:count)
             statuses += Termtter::API.twitter.list_statuses(user_name, slug, options)
           else
-            user_name = normalize_as_user_name(user.sub(/\/$/, ''))
-            statuses += Termtter::API.twitter.user_timeline(user_name, options)
+            begin
+              if user =~ /^\d+$/
+                profile = Termtter::API.twitter.user(nil, :screen_name => user) rescue nil
+                unless profile
+                  status  = Termtter::API.twitter.show(user) rescue nil
+                  user    = status.user.screen_name if status
+                end
+              end
+              user_name = normalize_as_user_name(user.sub(/\/$/, ''))
+              statuses += Termtter::API.twitter.user_timeline(user_name, options)
+            rescue Rubytter::APIError => e
+              last_error = e
+            end
           end
         end
       end
       output(statuses, event)
+      raise last_error if last_error
     },
     :help => ["list,l [USERNAME]/[SLUG] [-COUNT]", "List the posts"]
   )
@@ -56,15 +71,17 @@ module Termtter::Client
       # TODO: show more information of lists
       lists = Termtter::API.twitter.lists(user_name).lists
       public_storage[:lists] += lists.map(&:full_name)
-      puts lists.map{|i| i.full_name}
+      puts lists.map{|i| i.full_name}.join("\n")
     },
-    :help => ["lists [USERNAME]", "Show Lists"]
+    :help => ["list list [USERNAME]", "Show Lists"]
   )
 
   register_command(
     :name => %s{list follow},
+    :alias => %s{list add},
     :exec => lambda { |arg|
-      slug, *users = arg.split(' ')
+      list_name, *users = arg.split(' ')
+      slug = list_name_to_slug(list_name)
       users.each{ |screen_name|
         begin
           user = Termtter::API.twitter.cached_user(screen_name) || Termtter::API.twitter.user(screen_name)
@@ -75,13 +92,14 @@ module Termtter::Client
         end
       }
     },
-    :help => ["list follow SLUG USERNAME", "Follow users to the list"]
+    :help => ["list follow|add LISTNAME USERNAME", "Follow users to the list"]
     )
 
   register_command(
     :name => %s{list remove},
     :exec => lambda { |arg|
-      slug, *users = arg.split(' ')
+      list_name, *users = arg.split(' ')
+      slug = list_name_to_slug(list_name)
       users.each{ |screen_name|
         begin
           user = Termtter::API.twitter.cached_user(screen_name) || Termtter::API.twitter.user(screen_name)
@@ -92,7 +110,7 @@ module Termtter::Client
         end
       }
     },
-    :help => ["list remove SLUG USERNAME", "Remove user(s) from the list"]
+    :help => ["list remove LISTNAME USERNAME", "Remove user(s) from the list"]
     )
 
   register_command(
@@ -107,6 +125,7 @@ module Termtter::Client
         opt.parse(options)
       }
       list = Termtter::API.twitter.create_list(slug, param)
+      public_storage[:lists] << list.full_name
       p [list.full_name, param]
     },
     :help => ["list create SLUG [--description VALUE] [--private]", "Create list"]
@@ -115,15 +134,40 @@ module Termtter::Client
   register_command(
     :name => %s{list delete},
     :exec => lambda { |arg|
-      arg.split(' ').each{ |slug|
+      return unless confirm("Are you sure?")
+      arg.split(' ').each{ |list_name|
         begin
+          slug = list_name_to_slug(list_name)
           list = Termtter::API.twitter.delete_list(slug)
+          public_storage[:lists].delete(list.full_name)
           puts "#{list.full_name} deleted"
         rescue => e
           handle_error(e)
         end
       }
     },
-    :help => ["list delete SLUG", "Delete list"]
-    )
+    :help => ["list delete LISTNAME", "Delete list"]
+  )
+
+  register_command(
+    :name => %s{list show},
+    :exec => lambda { |arg|
+      raise ArgumentError unless /([^\s]*\/[^\s]+)/ =~ arg
+      user_name, slug = *arg.split('/')
+      user_name = config.user_name if user_name.empty?
+      user_name = normalize_as_user_name(user_name)
+      list = Termtter::API.twitter.list(user_name, slug)
+      attrs = %w[ full_name slug description mode id member_count subscriber_count]
+      label_width = attrs.map(&:size).max
+      attrs.each do |attr|
+        value = list.__send__(attr.to_sym)
+        puts "#{attr.gsub('_', ' ').rjust(label_width)}: #{value}"
+      end
+    },
+    :help => ["list show LISTNAME", "Show the detail of list"]
+  )
+
+  def self.list_name_to_slug(list_name)
+    list_name[/([^\/]*)$/]
+  end
 end
